Lift a formula into a formula function in one place
`invariant!` builds its `#[thrust::formula_fn]` from a closure and the context the
closure was written in: the in-scope generics are re-declared on the function and
instantiated via turbofish, `Self` becomes the impl's self type or a synthetic type
parameter in a trait, and the receiver `self` becomes a `__thrust_self` parameter.
None of that is particular to an invariant -- it is what any formula written inside a
function body needs to survive being lifted out of it.

Move it to `formula_fn_lifting`, which takes the parameters and the body and returns
the item plus the expression naming it, and leave `invariant` with the part that is
its own: reading the closure and emitting the marker call.
